+++ app/controllers/staff/planned_disbursements_controller.rb
@@ -53,7 +53,6 @@ class Staff::PlannedDisbursementsController < Staff::BaseController
 
   private def planned_disbursement_params
     params.require(:planned_disbursement).permit(
-      :planned_disbursement_type,
       :currency,
       :value,
       :providing_organisation_name,