+++ spec/models/planned_disbursement_spec.rb
@@ -42,14 +42,4 @@ RSpec.describe PlannedDisbursement, type: :model do
     it { should strip_attribute(:providing_organisation_reference) }
     it { should strip_attribute(:receiving_organisation_reference) }
   end
-
-  describe "validations" do
-    context "when the planned_disbursement_type is blank" do
-      it "displays the appropriate error message" do
-        planned_disbursement = build(:planned_disbursement, planned_disbursement_type: nil)
-        expect(planned_disbursement.valid?).to be_falsey
-        expect(planned_disbursement.errors[:planned_disbursement_type]).to include t("activerecord.errors.models.planned_disbursement.attributes.planned_disbursement_type.blank")
-      end
-    end
-  end
 end